+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/WorkflowBulk/workflow-bulk.js
