+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/WorkflowExec.css
