+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/details-modal/details-modal.tsx
